pyproject.toml
[project]
name = "pyke-lol"
version = "3.0.0"
description = "A Riot API wrapper specifically for League of Legends."
authors = [{ name = "diodemusic" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["httpx>=0.28.1,<0.29.0"]
keywords = [
    "riot api",
    "league of legends",
    "lol api",
    "riot python",
    "riot wrapper",
    "LoL",
    "League of Legends",
    "Riot Games",
    "API",
    "REST",
    "restful api",
]

[project.optional-dependencies]
dev = [
    "pytest-asyncio>=1.3.0,<2.0.0",
    "anyio>=4.12.1,<5.0.0",
    "respx>=0.22.0,<0.23.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "black>=24.0.0",
    "mypy>=1.18.2,<2.0.0",
    "bandit>=1.8.6,<2.0.0",
    "pdoc>=15.0.4,<16.0.0",
    "datamodel-code-generator>=0.34.0,<0.35.0",
    "case-converter>=1.2.0,<2.0.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "pyke", from = "src" }]

[tool.pyright]
reportPrivateUsage = false
typeCheckingMode = "strict"

[tool.ruff.per-file-ignores]
"tests/*" = ["F401", "F811"]

[[tool.pyright.executionEnvironments]]
root = "tests"
reportUnusedImport = "none"
reportMissingTypeStubs = "none"

exclude = [".env", "tests/*"]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s] %(name)s: %(message)s"
log_cli_date_format = "%H:%M:%S"
